tow sum

+++ array/majority-element.go
@@ -4,10 +4,10 @@ func majorityElement(nums []int) int {
 
 	m := map[int]int{}
 
-	l := float64(len(nums)) / 2
+	l := float32(len(nums)) / 2
 	for _, v := range nums {
 		m[v]++
-		if float64(m[v]) > l {
+		if float32(m[v]) > l {
 			return v
 		}
 	}
